Integrate retail-executable-spec, session, content changes
Curated integration snapshot combining 6 adjacent ancestry steps.

- PLAN_19 §8: log RWU-19-198; dispatch WU-19-213
- PLAN_19 §8: log RWU-19-199; dispatch WU-19-214
- PLAN_19 §8: log WU-19-212

Intermediate snapshots retain historical development state; they are not individually certified builds.

+++ internal/content/compile_weapon.go
@@ -32,10 +32,10 @@ type WeaponDef struct {
 	// itself is Established for all nine — "a 16-bit store then wraps the
 	// truncated value modulo 65,536" [06 §7.3] — and only the *extension*
 	// each reader applies to the stored word separates the signed from the
-	// unsigned form. Six of the nine name their reader's extension and are
-	// wrapped below; three (BurstRate, Duration, SmokeDelay) do not, and are
-	// left as the plain *30 truncation with an open-question marker at the compile
-	// site rather than guessed. Measured: over the 77-file, 198-section stock
+	// unsigned form. All nine now name their reader's extension and are
+	// wrapped below: the last three (BurstRate, Duration, SmokeDelay) were
+	// traced in RWU-19-198 and are zero-extended by every reader
+	// [06 R-WPN-05 §12]. Measured: over the 77-file, 198-section stock
 	// weapon family every one of the nine compiles to a tick count inside
 	// 0..32767, where all three forms agree, so no stock weapon can tell them
 	// apart (WU-19-167 census).
@@ -44,10 +44,10 @@ type WeaponDef struct {
 	WeaponAcceleration int32   // weaponacceleration *65536/900 truncated [02 "Weapon record"]
 	ReloadTime         int32   // reloadtime *30 truncated ticks, wrapped to a signed 16-bit store [06 §4.2]
 	WeaponTimer        int32   // weapontimer *30 truncated, wrapped to an unsigned 16-bit store [06 §7.3]
-	BurstRate          int32   // burstrate *30 truncated [02 "Weapon record"]; 16-bit store established [06 §7.3], reader extension is not — TODO(question)
-	Duration           int32   // duration *30 truncated [02 "Weapon record"]; 16-bit store established [06 §7.3], reader extension is not — TODO(question)
+	BurstRate          int32   // burstrate *30 truncated, wrapped to an unsigned 16-bit store [02 R-KEYS-01 §6] [06 R-WPN-05 §12]
+	Duration           int32   // duration *30 truncated, wrapped to an unsigned 16-bit store [02 R-KEYS-01 §6] [06 R-WPN-05 §12]
 	RandomDecay        int32   // randomdecay *30 truncated, wrapped to an unsigned 16-bit store [06 §4.3]
-	SmokeDelay         int32   // smokedelay *30 truncated [02 "Weapon record"]; 16-bit store established [06 §7.3], reader extension is not — TODO(question)
+	SmokeDelay         int32   // smokedelay *30 truncated, wrapped to an unsigned 16-bit store [02 R-KEYS-01 §6] [06 R-WPN-05 §12]
 	FlightTime         int32   // flighttime *30 truncated, wrapped to an unsigned 16-bit store [06 §6.6]
 	HoldTime           int32   // holdtime *30 truncated, wrapped to a signed 16-bit store [07 "in-flight camera move"]
 	ShakeDuration      int32   // shakeduration *30 truncated; established as a 32-bit store, no further truncation [02 R-KEYS-01 §2]
@@ -212,36 +212,25 @@ func compileWeaponSection(section *formats.Section, sectionName string, prov Pro
 	// here with `int32(int16(...))` (sign-extended back) or
 	// `int32(uint16(...))` (zero-extended back).
 	//
-	// The remaining three (burstrate, duration, smokedelay) name no reader
-	// extension anywhere in research/retail-executable-spec, so they are left
-	// as a plain *30 truncation with no second wrap; choosing a sign here
-	// would be the invented-constant failure mode this codebase forbids. The
-	// three arms disagree only outside 0..32767, and a census of the stock
-	// weapon family (77 files, 198 sections) finds every one of the nine keys
-	// inside that range in every section, so nothing shipped distinguishes
-	// them (WU-19-167).
+	// The remaining three (burstrate, duration, smokedelay) used to be left as
+	// a plain *30 truncation because no reader extension was written up. It
+	// now is: every reader of the three words zero-extends them and compares
+	// unsigned — the burst scheduler's due test, its ">4" refresh gate and its
+	// deadline advance, the beam latch, and the trail-smoke deadline
+	// [06 R-WPN-05 §12] [02 R-KEYS-01 §6] — so they take the same unsigned wrap
+	// as weapontimer. The three arms disagree only outside 0..32767, and a
+	// census of the stock weapon family (77 files, 198 sections) finds every
+	// one of the nine keys inside that range in every section, so nothing
+	// shipped distinguishes them (WU-19-167); third-party content can.
 	reloadTime := int32(int16(int32(section.FloatValue("reloadtime", 0) * 30.0)))    // signed 16-bit store [06 §4.2]
 	weaponTimer := int32(uint16(int32(section.FloatValue("weapontimer", 0) * 30.0))) // unsigned 16-bit store [06 §7.3]
-	// TODO(question): burstrate's 16-bit store is Established [06 §7.3]; the extension its readers
-	// apply to that word is not. [06 §4.3] casts the burst deadline SUM to unsigned
-	// (`(uint32)(creationTick + burstrate) <= currentTick`) and tests the interval against four
-	// ("strictly greater than four"), but neither states how the stored word widens before either
-	// use. Decider: the burst scheduler's two burstrate loads, written up in [06 §4.3].
-	burstRate := int32(section.FloatValue("burstrate", 0) * 30.0)
-	// TODO(question): duration's (beamweapon) 16-bit store is Established [06 §7.3]; the extension its
-	// one reader applies is not. That reader is the beam latch of [06 §6.10],
-	// `creationTick + duration < currentTick`, whose comparison strictness is Established while the
-	// addend's widening is not. Decider: the latch's duration load, written up in [06 §6.10].
-	duration := int32(section.FloatValue("duration", 0) * 30.0)
+	burstRate := int32(uint16(int32(section.FloatValue("burstrate", 0) * 30.0)))     // unsigned 16-bit store, zero-extended by all three burst-scheduler loads [06 R-WPN-05 §12]
+	duration := int32(uint16(int32(section.FloatValue("duration", 0) * 30.0)))       // unsigned 16-bit store, zero-extended by the beam latch [06 R-WPN-05 §12]
 	randomDecay := int32(uint16(int32(section.FloatValue("randomdecay", 0) * 30.0))) // unsigned 16-bit store [06 §4.3] ("an unsigned 16-bit shift")
-	// TODO(question): smokedelay's 16-bit store is Established [06 §7.3]; the extension its one reader
-	// applies is not. That reader is the additive trail-smoke deadline of [06 §7.3]
-	// (`smokeDeadline += smokedelay`), which the section states additively but without the widening.
-	// Decider: the trail-smoke deadline's smokedelay load, written up in [06 §7.3].
-	smokeDelay := int32(section.FloatValue("smokedelay", 0) * 30.0)
-	flightTime := int32(uint16(int32(section.FloatValue("flighttime", 0) * 30.0))) // unsigned 16-bit store [06 §6.6] ("(uint16)flighttime")
-	holdTime := int32(int16(int32(section.FloatValue("holdtime", 0) * 30.0)))      // signed 16-bit store [07 "in-flight camera move"] ("the count is signed")
-	shakeDuration := int32(section.FloatValue("shakeduration", 0) * 30.0)          // established 32-bit store [02 R-KEYS-01 §2], no further truncation
+	smokeDelay := int32(uint16(int32(section.FloatValue("smokedelay", 0) * 30.0)))   // unsigned 16-bit store, zero-extended by the trail-smoke deadline [06 R-WPN-05 §12]
+	flightTime := int32(uint16(int32(section.FloatValue("flighttime", 0) * 30.0)))   // unsigned 16-bit store [06 §6.6] ("(uint16)flighttime")
+	holdTime := int32(int16(int32(section.FloatValue("holdtime", 0) * 30.0)))        // signed 16-bit store [07 "in-flight camera move"] ("the count is signed")
+	shakeDuration := int32(section.FloatValue("shakeduration", 0) * 30.0)            // established 32-bit store [02 R-KEYS-01 §2], no further truncation
 	// turnrate *1/30 truncated per tick, wrapped to an unsigned 16-bit store [06 §6.7] ("zero-extended from its 16-bit store")
 	turnRate := int32(uint16(int32(section.FloatValue("turnrate", 0) * (1.0 / 30.0))))
 	// minbarrelangle *pi/180 radians default -11.25 — composed exactly as
